organized pg files and wrote instructions for pg config
